UnwantedImageDetection_server/requirements.txt
opencv-python
numpy
pycuda
torch
torchvision
fastapi
aiohttp
uvicorn
python-dotenv